silence -Wsign-compare warning with explicit cast

+++ include/process.h
@@ -281,7 +281,7 @@ class pipe_ostreambuf : public std::streambuf
             // move the put_back area to the front
             const auto dest = base;
             const auto src  = egptr() - put_back_size_ < dest ? dest : egptr() - put_back_size_;
-            const auto area = egptr() - dest < put_back_size_ ? egptr() - dest : put_back_size_;
+            const auto area = static_cast<std::size_t>(egptr() - dest) < put_back_size_ ? egptr() - dest : put_back_size_;
             std::memmove(dest, src, area);
             start += put_back_size_;
         }
